Fix multi-FramePtr phi-jump issue
Summary:
In optimized translations, we'll generate instructions like "Jmp t635:FramePtr, t645:FramePtr -> B535<Unlikely>" where we phi multiple FramePtrs. In particular, we'll do so when we guard multiple values while inlining multiple levels deep, because we share a single ReqBindJmp merge point for all the guards.

This instruction is kind of weird, since the phijmp's dst defines the rvmfp() physical register twice. vasm-graph-color makes an exception for it. However, we'll crash if we later eliminate all but one of the guards, because we'll turn the phijmp into a copyargs and assert that it defines rvmfp() twice. We just have to fix the assertion.

In the new unit test, we hit this edge case by inlining two frames and then guarding the result of "mixed_value". The more likely tracelet guards to InitNull; the less likely one guards to Arr and then to Arr=PackedKind, but we later eliminate the Arr guard using the RepoAuthType. (The "darray" option is needed so that the RepoAuthType is ?Arr instead of ?Varr.)

Reviewed By: ricklavoie

Differential Revision: D20322192

fbshipit-source-id: a70ebe5130a2d8ae0a6b9ab7fe333f9b242ab44e