Introduce arena-compatible Reason type
Summary:
This diff introduces an arena-compatible Reason type.

Note that this is the first time we diverge from the OCaml in a somewhat significant way.

We'll have:

```
pub struct PReason_<'a> {
    pub pos: Option<&'a Pos>,
    pub reason: Reason<'a>,
}

pub type PReason<'a> = &'a PReason_<'a>;

pub enum Reason<'a> {
    Rnone,
    Rwitness,
    /// Used as an index into a vector-like
    /// array or string. Position of indexing,
    /// reason for the indexed type
    Ridx(&'a Reason<'a>),
    [...]
}
```

Note the addition of the `mk_rnone` method to the `TypeBuilder` struct. It is sadly already becoming a bit of a kitchen sink, but I don't think that's necessarily bad: in the end we will probably be allocating at most 3 different types in the arena:

1. Typs
2. Reasons
3. Map nodes

So it will be limited. We might want to rename it in the future though.

Reviewed By: andrewjkennedy

Differential Revision: D20304717

fbshipit-source-id: 61967f1bd7099c81230ad04747cb1495d7cb2b41