Clean up decreflocals in unwindVM
Summary: Long time ago, on D1384240, decref locals logic in unwindVM/Php was made to be a lambda where if you have a resumed async function that is failed, and you're unwinding this function, we needed to avoid decreffing its locals. This is no longer the case as we can no longer have resumed async functions that failed on the rbp chain. Lets clean up this logic and assert that it cannot happen.

Reviewed By: markw65

Differential Revision: D20287874

fbshipit-source-id: 401c91c77ff0da137ee5a67d148cf4a7655b558d