Fix range-formatting bug
Summary: A bug in `text_with_formatted_ranges` resulted in dropping the end of files when we invoke this helper with an empty list of formatted ranges (which occurred in partially-generated files when every manual section was empty). We should copy characters all the way until the end_offset rather than until we've hit the length of the range in the original source text.

Reviewed By: arxanas

Differential Revision: D20318830

fbshipit-source-id: 4b78a870db9894ce04ce4afaa1ab4173f3ddc1af