Correctly handle whitespace when formatted range boundaries fall at a line boundary rather than an atom boundary
Summary:
Since {D5843678}, we have had the guarantee that the intervals being formatted in --diff mode begin and end at the boundaries of atoms. An atom is an indivisible chunk of text in hackfmt. Often there is a 1:1 correspondence between atoms and tokens, but some tokens are subdivided into smaller atoms (for example, a multiline string is a single token, but is split into atoms on newlines).

This invariant was broken by {D20124500}. Noformat ranges begin and end at line boundaries. By subtracting them from our diff-formatting ranges, we ended up with formatting ranges which may begin or end either at a line boundary or an atom boundary, but we still handled them as if they all fell at atom boundaries.

Instead, for each range boundary, we should determine whether it falls at an atom boundary. If so, we omit surrounding whitespace (opting into the behavior described in the summary of {D5843678}). If not, we print surrounding whitespace, since the surrounding whitespace associated with a chunk is designed to be correct for line-boundary-based range-formatting.

I also replaced the Line_splitter.print with a Line_splitter.solve here. Computing all the solve states ahead of time was a performance optimization which probably isn't necessary, and besides, it prevents us from using the range behavior of Line_splitter.solve (first implemented in {D5853970}). This range-formatting behavior uses the existing line breaks outside the formatting range to produce a formatting solution for the range that is more consistent with the unformatted context. FIXME-based formatting suppression makes this even more critical. Without it, in the added test cases, hackfmt will want to collapse `vec[1, 2, 3]` onto one line, producing odd-looking results like this:

```
    /* HH_FIXME[4110] */
    return vec[1,
2, 3];
```

Reviewed By: arxanas

Differential Revision: D20318829

fbshipit-source-id: be3380c8faf3f036d74a4c91566d48e371f50f56