Disambiguate array kind/array type checks
Summary:
We're planning on breaking away from the current world where array
kind can be used to infer both the physical layout of an ArrayData and the
DataType. So get rid of the `ArrayData::is$X` methods and replace them with
`ArrayData::is${X}Kind` and `ArrayData::is${X}Type`.

The isKind flavors simply report the array kind like the old fns do; the isType
flavors report on the required DataType of the array.

Anywhere that uses these fns to guard on the physical layout of the array gets
an isKind check--the others get an isType check.

Reviewed By: billf

Differential Revision: D19958200

fbshipit-source-id: 54a3433dc191c104ce34700cf27c1f1c66a5bce6