Inject logging into magic methods
Summary: For `__get/set/isset/unset`, inject the call within hhvm. For __call, there are certain situations where it is triggered where "we are in the middle of the opcode and stack is inconsistent and we are not able to throw anything" - Jan, so inject the logging manually in the methods themselves.

Reviewed By: fahndrich

Differential Revision: D20313715

fbshipit-source-id: 61d4b91eacc1b89d541071d11315d8ec0bccbed1