Don't set_global_context around process_changed_file
Summary:
This call to set_global_context was probably a historical legacy from a time when process_changed_files didn't take `ctx`.

The effect of calling set_global_context is solely to prevent any calls to `Provider_utils.compute_tast_but_respect_quarantine` from quarantining. The presence of a ctx was enough to stop that from happening.

This effect isn't something that's appropriate for process_changed_files. So, I'm removing it.

Reviewed By: arxanas

Differential Revision: D20347035

fbshipit-source-id: 8f0e0c60715681153d8b70ad2afbef22183b923e