Kill source mapping option across stack
Summary:
This option is explicitly set to true at every entry point to HackC,
as well as in tests, but it's true by default in HHVM.
It's also an outlier because it has 2 aliases when used as CLI argument:
- hack.compiler.sourcemapping
- eval.disassemblersourcemapping

neither of which matches its JSON key: eval.disassembler_source_mapping.

All of the abovee adds significant complexity, so kill it altogether.

Reviewed By: shiqicao, alexeyt

Differential Revision: D20295585

fbshipit-source-id: bed0040227143b6df122e2f11545bde909ace70e