Update to Rust 1.42.0
Summary:
This diff upgrades fbcode platform007 and platform009 to rust 1.42.0 + a cherry-pick of https://github.com/rust-lang/rust/pull/69685 for AddressSanitizer support, and xplat to plain 1.42.0.

Be aware that the xplat toolchain, for which we use the official upstream releases and don't build from source, no longer supports armv7-apple-ios and armv7s-apple-ios targets as of 1.42.0.

Differential Revision: D20379253

fbshipit-source-id: c98fecc35dbe077f8dd1b8c7a7bab098795ac2dd