Make FileAwait take std::chrono::nanoseconds, add _OS\poll_async()
Summary:
- add `poll_async()`, which is like `stream_await()` except:
  - takes an HSL FileDescriptor rather than a PHP resource
  - takes an int ns timeout instead of a double second timeout
- make FileAwait take an int timeout as:
  - we end up converting to an int anyway
  - the double behavior has led to bugs in the past - e.g.
    timeouts > 0.0 < 0.0001 become an ms timeout of 0, which
    means 'no timeout'
- moved the double -> milliseconds part to File::await for backwards compatibility
- added `_OS\pipe()` to make this easy to test.

Using nanoseconds as much as possible as this is usually the kernel granularity; converting to ms with ceil as folly does this internally if using higher-resolution - would rather make an explicit choice in HHVM if we change it in the future.

Instead of special-casing the `0.0001` -> `0` case, always use `ceil()` as timeouts are always a lower bound anyway.

Not sure this will be publicly exposed; if so, will likely use enum
return and parameter types in the non-private wrapper.

Reviewed By: jano

Differential Revision: D20230942

fbshipit-source-id: ceae54e0d5fc8770af4006f80467030d86e188d8