Do not depend on loop order for solving to a shallow match
Summary:
D20374080 fixed an issue where we lost some constraints when we partially solved a type variable by finding a shallow match between the lower and upper bounds. The fix as written can still potentially lose constraints depending on the order we iterate through the lower and upper bounds.

Consider the case when we have the following bounds
```
Vector<int> <: v
ConstVector<string> <: v
v <: ConstVector<arraykey>
```
When we solve this using a shallow-match then we bind `v := ConstVector<v1>` and we do:
```
ConstVector<string> <: ConstVector<v1>
ConstVector<v1> <: ConstVector<arraykey>
```
It is clear we need to also do:
```
Vector<int> <: ConstVector<v1>
```
but since we add the sub type constraints in the same fold that we bind it is possible we won't do the check. For instance if the iteration order was:

```
(Vector<int>, ConstVector<arraykey>)
(ConstVector<string>, ConstVector<arraykey>)
```
then we won't do a shallow match on the Vector/ConstVector pair, but will for the second but never go back to add the additional constraint.

To make this more obviously correct I propose rewriting the code so the search for a shallow match is done separately from applying the needed constraints.

The approach can thus be described as follows:
1. For a variable `v`, look for a lower bound `L` that is a shallow match of an upper bound `U`.
2. Freshen the types in `L`. Call this new type `match_ty`
3. Compute `v <: match_ty` and `match_ty <: v`. This will utilize our subtyping code to ensure `match_ty` has the proper constraints.
4. Bind `v := match_ty`

Reviewed By: mpu

Differential Revision: D20374080

fbshipit-source-id: f0d04c9ee6e74489fed43c96dcf6645cf3942a9f