Include repo-schema in mangled unit SHA
Summary:
Originally there was no need to include the repo-schema in
the mangled unit SHA because we only looked up units on a table keyed
by the repo-schema. Now that we're using the mangled unit SHA in more
contexts (IE, bccache), it's safer to include the repo-schema, to
ensure we don't accidently re-use units across different compiler
versions.

Reviewed By: alexeyt

Differential Revision: D20393075

fbshipit-source-id: c08a4cf004de9810cca7aedb03484c4bd610be56