Pre-emptively fix/delete tests implicitly casting non-arraykey keys for arrays 1/n
Summary:
Future changes will turn on enforcement of the `HackArrCompatCheckArrayKeyCast`  option, making it such that any non-arraykey key used with an array type will start throwing. This is how hack arrays have behaved from the start, and it is time to bring legacy arrays to parity.

Because this is such a widespread PHPism, there's a hundred and some change tests that need fixing and I'd rather not do that in one gigantic diff, so I'm pre-emptively dealing with tests in chunks prior to making the switch.

The test `test/slow/hack_arr_compat/implicit-key.php` is here to stay and provides comprehensive coverage for this coercion behavior so I'm erring on the side of deletion where possible, but do advise which tests are best updated instead of deleted.

Reviewed By: dneiter

Differential Revision: D20231771

fbshipit-source-id: 6a9884e53b01b18d277623965fe0bb41f699fca7