][ctx][easy] Move AST-computing methods from `Provider_utils` to `Ast_provider`
Summary: The `Naming_provider` needs to parse files in the context to find their symbols. We could do this using the methods already in `Ast_provider`, but that means that we would force a re-parse for every file in the context whenever we looked at it (potentially once per name lookup)! To take advantage of caching behavior, move the cache-aware functions from `Provider_utils` into `Ast_provider`.

Reviewed By: ljw1004

Differential Revision: D20425805

fbshipit-source-id: 214a1986e9259e19da91e304b1c56a66ddb7b972