Prototype typed AST, decl provider, typing traversal
Summary:
As in title. This is enough logic to produce and print typed AST for following file:
```
<?hh

function g(): int {}

function f(): void {
  g();
}
```
Where internal `f` expression is annotated with `int` (other types and reasons are dummy).

The traversal in `typing.rs` is not mirroring the one in `typing.ml` yet, because there is too much complication there - IMHO it's better to get down to expressions as ASAP and start work from there. The individual parts are:

- `tast.rs` is (similarly to `ast.rs`) a set of aliases on `aast.rs`
- `decl_provider.rs` is the trait object for decl providers
- for the purposes of `hh_check`, `DeclProvider` is implemented by `TestDeclProvider`
- `TypeBuilder` and `DeclProvider` are carried around in enviroment
- `typing.rs` is the main traversal routine
- `Debug` is added to a bunch of types to be able to print TAST
- some `*` imports are replaced with more targeted ones, because various versions of `Ty_` collide too much (in general, I feel there is too much of `*` and `pub use`, and sometimes very unexpected symbols end up in current scope)
- some naming cleanup in `Cargo` files

Next step: try to print the `TAST` with OCaml printer instead of Rust `Debug`.

Reviewed By: shiqicao

Differential Revision: D20376295

fbshipit-source-id: 0be5394af5f35c10560ad853d2e4a1047d44de27