Eliminate generic LdIterBase
Summary:
We'd like to relax the constraint that HeaderKind implies DataType. This constraint is annoying both for bespoke types and for KindOfDVArray specialization:
- For bespoke types, we might want one HeaderKind ("bespoke") to map to multiple array-like data types.
- For KindOfDVArray, we want this mapping to depend on runtime flags.

We only rely on this constraint in one place: to implement a generic load of an iterator base (where we don't already know its type). This load is slow, so we only use it to profile this base in case we end up specializing the iterator later. This profiling doesn't help much - it only kicks in one time at the end of the loop, not once per iteration. Delete it.

Reviewed By: ricklavoie

Differential Revision: D20400183

fbshipit-source-id: 8f96e33b01a3812bb6c13e0f36848644e1c093f2