`hh --identify` [1/3] plumbing
Summary:
This adds a new mode, `hh_client check --identify <symbol>` which identifies whether the symbol is defined. If so, it returns the symbol's full name, kind and location. If the symbol is defined more than once, it returns all of them.

We already had `--identify-function` which takes line+column and returns the definition of the symbol at that location. This new `--identify` is similar to that, but uses name rather than location.

I think that this is a reasonable feature to add, and a reasonable one to exist in the hh_server architecture. In that world it will be fulfilled straightforwardly by the decl service with no additional processing needed.

I made things simple. It only emits json. It only emits a small number of fields. That's so we don't tie our hands in future.

Reviewed By: arxanas

Differential Revision: D20396262

fbshipit-source-id: cf81d835607fdc6843c0af3d41fa1a489792b58f