Add gdb frame filter and decorator
Summary:
Once we install our unwinder, `bt` can unwind properly through jitted
code, but it still doesn't know anything about the jitted frames, so
we have our own walkstk that mostly emulates bt, but also adds
function names, file names and (usually incorrect) line numbers to the
jitted frames.

One downside of this is that it matched the bt format when it was
written, but doesn't match as the bt format evolves - such as the
recently added colorization.

I started looking through the gdb python unwinder docs again to see if
we could add anything to describe the frames, and it turns out that
there's nothing more we can do there. However, there are frame
filters, and frame decorators which can be applied to any
frames. Filters can decide whether to display frames at all, and which
decorators to use on the ones that are shown, and decorators determine
what the frame actually looks like.

So I wrote a filter to identify jitted frames, and a decorator to
display them properly. The result is that the output of bt now looks a
lot like the output of walkstk except:
 - function names include their parameter names `foo(bar, baz)` instead of `foo`
 - argument values are displayed after the function name
 - line numbers are displayed based on the correct pc
 - local variables are also displayed with `bt full`

We still had the issue that none of this started up automatically (so
you either had to explicitly `unwinder init` or `walkstk` to
initialize the unwinder). The problem there was that we can't
initialize the unwinder until hhvm is loaded, and we're either
debugging a live inferior, or have a core loaded. So I added an event
listener that tries to initialize the unwinder just before displaying
the prompt.

Example backtrace:

{P127428502}

Reviewed By: oulgen

Differential Revision: D20428552

fbshipit-source-id: 07cadef91ba9dcd1e24cbaeae84d6c76b8c0bdc1