Move `add_entry` functions from `Provider_utils` to `Provider_context`
Summary: It used to be that we eagerly parsed the file when constructing a new entry, so we had to put the entry constructor inside something higher in the dependency graph than `Provider_context`. That's no longer the case, so we can put the entry constructor back in `Provider_context`.

Reviewed By: ljw1004

Differential Revision: D20425804

fbshipit-source-id: 9f544e2c8bc754ebfe1927ce40452f16861d3591