Raise Failure rather than RustException on panic
Summary:
This diff removes the `RustException` error class, with id
`"rust exception"`, that was previously used to pass errors across
the ffi boundary between ocaml and Rust when Rust panics.

Why make this change? Because `ocaml::core::fail::fail_with` raises
an ocaml error of type `Failure`, which is the standard way to
indicate unrecoverable exceptions in ocaml (and a Rust panic is not
recoverable)

In addition to being more standard it reduces the
chances of error - we no longer need to register `"rust exception"`
as an exception type key in ocaml, and we sidestep a bug we've
encountered where printing the custom RustException could result
in null bytes, which led to Utf-8 errors in Rust in certain cases
where Rust was calling ocaml which called back into Rust.

Differential Revision: D20435202

fbshipit-source-id: accfd23d33670d24feb33d1a4518b5e546d60150