Add skeleton struct for RFuncData
Summary:
Goal is to add support for first class function pointers that also support reified generics:
```
$x = foo<int, string, C>;
$x($args);
```
Plan is to add a new heap tracked typed value called RFunc to represent a combo of the func pointer and the reified generics.

This diff is only the skeleton of RFunc. The next diff fills it out some more with a constructor, etc.

Reviewed By: oulgen

Differential Revision: D19827729

fbshipit-source-id: a86ba58673eee01ea401a79a778ab4d0f0c6c4ba