Implement methods of RFunc
Summary:
Implement a constructor and destructor for RFunc.

In order to make it a heap request allocated object, it seems I need to create add a new HeaderKind. I've left some of those blank because I'm not sure what to do / what the functions are doing with regards to heap request allocated objects.

Reviewed By: oulgen

Differential Revision: D19885448

fbshipit-source-id: d96374b7feeb4d71e702ac8756236d16804be038