Option to change fuzzy comparisons between Hack arrays and boolean/null
Summary:
As mentioned in the [HHVM 4.45 release notes](https://hhvm.com/blog/2020/02/19/hhvm-4.45.html), comparing Hack arrays to `true`, `false`, and `null` with "fuzzy" equality comparison operators will be changed so that non-empty Hack arrays are `== true` and empty Hack arrays are `== false` and `== null`.

Setting `Eval.HackArrEmptyBasedBoolEqCmp=true` will allow opt-in to this new behavior, which makes it consistent with the existing behavior for other types, such as legacy (PHP) arrays.

Reviewed By: viratyosin

Differential Revision: D20401536

fbshipit-source-id: 1dacea2626ad45f68b73cfef116a1120d04c8fad