Do not use Ast_provider in Reified_check
Summary:
In {D18700484}, we began using Ast_provider in this TAST check to ban the use of class-level reified generics in static methods. This is an issue for the hh_server rearchitecture--we do not plan to provide TAST checks with an efficient way to fetch ASTs of other files.

I think we can get the same behavior just using Decl_provider, by looking up the static method as instantiated in its class-of-origin. This diff attempts to do so.

Reviewed By: Wilfred

Differential Revision: D20352111

fbshipit-source-id: 8fcaa2159a1c9938fa1b8f861b9d186a8feb03c8