Fix issues with global symbols in gdb-index
Summary:
I noticed that gdb wasn't finding certain global variables in certain
build modes. I initially thought it was a gdb bug, but then I tried
removing the index, and it worked.

Looking at a problem symbol in the index, I noticed that it was listed
as "static", rather than "global". Digging into the dwarf I found a
declaration that clearly marked it as external, followed by a
specification referring back to the declaration that did not.

The definition is supposed to inherit all of the "meaningful"
attributes (eg, clearly not the "declaration" attribute, or any
sibling attribute, for example) from its specification, but we weren't
doing that (except for the name).

This follows the specification and gathers attributes from there too.

Reviewed By: oulgen

Differential Revision: D20456813

fbshipit-source-id: 04f60f626c2a3e45b89d4245bc0b692a80c49421