only evict decls that are already there
Summary:
One of the root causes of T63645932 is that we called invalidate_context_decls at a time when ctx included a file which (1) didn't have an AST, (2) the underlying file no longer existed on disk.

We tried to invalidate all decls from that file. But first fetched it off disk to know which decls to evict. And got an exception while doing so.

This diff avoids that problem. It only invalidates decls for which we already have an AST.

Reviewed By: arxanas

Differential Revision: D20451749

fbshipit-source-id: cb5b4a200497b4c3a679c70dfdb2d0ebab5f7dd2