exn guarantee for respect_but_quarantine
Summary:
Previously this function didn't have clean exception semantics. In particular, if an exception arose while entering quarantine, then it would fail to leave the quarantine cleanly.

It's a bug (fixed next in the stack) that we failed to enter the quarantine cleanly. But nevertheless this function should be fixed.

Note that `Utils.with_context ~enter ~exit ~f` will decline to 'exit' if 'enter' raised an exception. That was what bit us. I've switched to `Utils.try_finally` which is less subtle.

I also added code to leave quarantine cleanly in the case where we only partially succeeded to enter quarantine.

Reviewed By: arxanas

Differential Revision: D20450313

fbshipit-source-id: b3c120e067f6f7e206f8909060c5fd4c1ded1258