retry HSL _OS\ functions on EINTR
Summary:
Was suggested to stop using exceptions for performance, given that EINTR and EWOULDBLOCK
are common.

Idea of this is
- stop EINTR being common
- EWOULDBLOCK can be avoided in library design via select or awaitable equivalents

Approach is to search both the Linux and BSD man pages for `EINTR`; if there are
warnings (like there are for `::close()`) leave alone; otherwise retry 5 times.

If there are 5 attempts, it probably actually is exceptional, so throw, like `close()` does

Reviewed By: paulbiss

Differential Revision: D20377643

fbshipit-source-id: fc122b4e3411b841769236667b40f0bbf9c3749a