Add context field to FCallArgs
Summary: `meth_caller()` needs to be able to call protected methods and properly resolve private methods in the correct context. To do that and preserve the use of top level functions as the implementation of `meth_caller()` we will need to allow call bytecodes to explicitly specify the context from which they are calling. We encode this as an optional field in `FCallArgs`

Reviewed By: ricklavoie

Differential Revision: D20329904

fbshipit-source-id: 95b06128556a4c775e517604175f830623f1e982