unrestrict-layout part 4: Eliminate append/remove in place
Summary:
As part of this stack I want to simplify the ArrayData interface methods. SetInPlace, RemoveInPlace, and AppendInPlace are all just like the non-in-place versions except that they don't do a COW check. Their main use is in the ArrayInit helpers. To fulfill that purpose, they can be specific methods on PackedArray and MixedArray - they don't need to be part of the general interface.

In this diff we drop RemoveInPlace and AppendInPlace, while still using PackedArray::AppendInPlace in VArrayInit and VecInit. I'll do SetInPlace next.

Reviewed By: ricklavoie

Differential Revision: D20472749

fbshipit-source-id: 1815fb7eb00c5b78284e20815ffabe5a16a6ed46