Fix crash in ThreadLocalFlat::onThreadExit
Summary:
Its possible for an onThreadExit function to be registered for a
ThreadLocalFlat even if the value is already destroyed. The easiest
way is if destroy() is called manually. This can cause a crash when
the thread is exiting because we'll call the destructor on garbage
memory. Add a check for the value already being destroyed.

Reviewed By: paulbiss

Differential Revision: D20479550

fbshipit-source-id: 3f34591c6872477a83c9d77661a9cdeb12e260c1