Unset maybe counted dead locals before a suspend
Summary:
In order to hold less memory while a functions frame is suspended, we can
unset locals that are no longer used, that may contain counted values.

This currently implements this during the DCE pass of HHBBC because we have the
liveness information there.  I am not the biggest fan of the exact current
setup, because it inserts instructions whose effects will not be fully
analyzed.  This isn't an issue today, but it will conflict with planned changes
I am working on to make locals share slots.

Differential Revision: D20490943

fbshipit-source-id: ef768edf5bf178e2b2ae8a3478ab6cb50cd0c4d9