Sandcastle utility
Summary:
Factoring out a utility module to invoke `scutil` in order to get the status and various timestamps of running jobs.

Besides the refactoring, the additional logic here is about reading the special environment file in order to find out the job (instance) ID that `hh_server` is executing in the context of, should it happen to be running in Sandcastle.

I named the module `Ci_util` because it may be used later in non-Facebook code. In the rest of my stack, I only log it in the remote worker initialization path, but it could also be logged in general.

Reviewed By: yinghuitan

Differential Revision: D19849718

fbshipit-source-id: a078a2ba4717a9c952a1e02aa09a752f68d81763