Fix func name display when m_fullName isn't set
Summary:
m_fullName is now computed lazily, but nameof wasn't
updated. Generally, we don't notice because when we crash, we attempt
to produce a php backtrace which means that all functions in the
backtrace get m_fullName set.

Sometimes generating the php backtrace fails, and sometimes we want to
see backtraces in other threads, so manually compute the full name if
it hasn't been cached.

Reviewed By: paulbiss

Differential Revision: D20492600

fbshipit-source-id: 9c77538c8dcd8831bf657d26c02555df70aaeb7d