RFC - invalidate_context_decls is only for local case
Summary:
It just kinda feels weird to me that invalidate_context_decls is only ever called for the local provider, but then it calls into invalidate_class which switches on whether it's the local or sharedmem provider.

This diff is to see what people think if we make it clear in the type system that invalidate_context_decls is only applicable for the shallow case.

Reviewed By: arxanas

Differential Revision: D20451790

fbshipit-source-id: 51dd36f7f727a00ddbe704eb160604297fff527f