unrestrict-layout part 6: Eliminate ArrayData::rvalStrict
Summary:
ArrayData::rvalStrict is just ArrayData::rval, except that if the value is missing and the array is a Hack array, it will raise one of a couple exceptions rather than returning a nullptr tv_rval. It's not worth implementing this method per-array-like; we already need to check for nullptr tv_rvals in ArrayData::get and in the JIT helpers (to convert them to ElemEmptyish for some minstr modes), so we might as well raise there too.

In this diff, we also drop the names "RvalInt" and "RvalStr" and standardize on "NvGetInt" and "NvGetStr". Here's why I chose "Nv" and not "Rval":
1. We hardcode the "Nv" name in a few critical places, notably BOLT inlining and in the assembly version of MixedArray::NvGetStr.
2. As we continue locking down these interfaces, we'll see that we can unify the lval and rval methods. At that point, the "Nv" name is a neutral one.

If we don't like this name, we can always change it in a follow-up - with a standard name it'll be easier to find and replace.

Reviewed By: jjgriego

Differential Revision: D20464208

fbshipit-source-id: 92e57d2a580f142f2245b0032893703791128184