unrestrict-layout part 5: Eliminate set in place
Summary:
...and here's the follow-up to delete SetInPlace. Again, the only significant user is ArrayInit, which always knows that the array is a mixed array.

This diff is a bit more involved than the previous one because we were calling the generic setInPlace in ArrayInit in a bunch of places that I had to switch to the MixedArray one. But that's also a good sign: if this change has any perf effect, it should be a small win, because we're de-virtualizing those calls.

Reviewed By: ricklavoie

Differential Revision: D20476851

fbshipit-source-id: 61798dc1534c061c143ba3f61c3538b7d95a8398