Fix job-queue thread flushing logic
Summary:
D20226679 changed the job-queue to eagerly flush when we are unable
to dequeue requests due to resource exhaustion.  It has a bug where
ableToDequeue is only set in the loop guard some of the time (only if it
doesn't short ciruit it).  Since ableToDequeue is used in the loop body, it may
be used before being set.  This could result in us always flushing thread
stacks when the job-queue is empty (basically all the time in normal
operation).

Differential Revision: D20512256

fbshipit-source-id: 6821ab930f3c002858feff08100aab6a326f7cbf