Thread `Provider_context.t` through `ServerFindLocals`
Summary:
`ServerFindLocals` is responsible for finding references to a local variable. This powers scenarios such as document-highlights and rename.

Currently, we do a re-parse of the file for every invocation. Document-highlights is one of the most frequent IDE actions taken (since it happens whenever you click anywhere in the document), so that's probably been adding several milliseconds of latency that keeps queueing up.

Switch this to use provider contexts so that we cache the AST that we're using between IDE requests for the same file.

Reviewed By: ljw1004

Differential Revision: D20518728

fbshipit-source-id: 548d3ac349eb2cf3c7f9f319ad3502dd59450d8c