Jit teardown for functions with all uncounted stack values and locals
Summary:
This diff is the first in the stack of diffs to jit function teardown. This particular diff implements stack teardown for functions with all stack values and locals definitely uncounted.
Since all the values are uncounted, all we need to do is to move the stack pointer (since there is no need for decreffing). This diff builds the scaffolding to make the make emitting decrefs trivial.

Reviewed By: markw65

Differential Revision: D19564399

fbshipit-source-id: 9817985dc4e85c3b55cdbec7c02539331f10eafb