omit empty heaps
Summary:
This is a tiny cosmetic change.

When getting telemetry in compute_tast about what's in memory, if any sharedmem heap has zero contents, then we don't even add it to the telemetry object. This makes the result easier to read.

In this diff, we extend the same courtesy to local_memory datastructures.

Reviewed By: arxanas

Differential Revision: D20552198

fbshipit-source-id: 9fcbbef5fb9c5c28fc8a62a6338525c3b424ccae