normalize status messages
Summary:
The status-bar is pretty small and was often getting truncated. Also, the status-bar has no indication of which language it is, so our message has to include it.
* I normalized status-bar messages to always start with "Hack:"
* I normalized to refer to Hack IDE services as just "Hack IDE" (abbreviated to "Hack:" in the status bar for reasons of space), and Hack Server as "hh_server"

My long-term vision is that the status bar will only ever be used to communicate the status of Hack IDE services, while the status of hh_server will be manifest from the Terminal>BuildTask window.

My medium-term vision on the way there is that hh_server status will be moved out of status bar, and into Diagnostics pane, similar to how Arc Lint puts up a diagnostic "hh_server isn't running hence diagnostics aren't available"

So, this diff is a step in that direction. It adopts the new terminology for the status-bar.

One final thing. I wanted to make the "OK" status indicator small, since it'll be there all the time and doesn't warrant much space in the 99% case where it's working. I picked a tick mark as the smallest thing I could fit. Java has been using a thumbs-up symbol, but I didn't like that because it showed up as a yellow emoji and I think that black-and-white is better and less conspicuous.

Reviewed By: arxanas

Differential Revision: D20484946

fbshipit-source-id: 2afcc7e9e7fcd4e41a7e2696aae202c367dd64b9