saved-state error messages 1/5 [API of saved_state_loader]
Summary:
I'm overhauling the way we show saved-state-load errors. The goal is to (1) reduce our support burden, (2) educate users so they can either self-solve, or know where to ask for help and what information to provide.

We have these UI surfaces:

* status-bar
* status-bar tooltip
* window/showMessage alert
* window/logMessage
* Hh_loggger.log

The status bar needs to be really short but also to point blame. I intend for this to have just the relevant keywords so a skilled user can recognize the issue.
* Hack: watchman fault
* Hack: www inconsistent

For the status-bar tooltip and the window/showMessage alert dialog, they can be just a little longer. The tooltip will wrap. The alert dialog by default truncates to just the first line of text, but ht euser can expand it. I intend for this to explain to the user the chain of blame and to point them further.
* A watchman fault stops Hack IDE from working. See Output>Hack for details.
* Hack IDE can't start because .hhconfig disagrees with your mergebase. See Output>Hack for details."

For the Output>Hack window, in previous diffs I've already ensured that it doesn't have spam. It only has actionable messages. The intent is that a user can read this and know what next steps they should take. That next step will either be running some commands themselves that they understand, or posting a request for help in an appropriate group and including enough details that a subject-matter-expert can help them.
* A watchman fault stops Hack IDE from working. This is not a bug in Hack or VSCode. Consider looking in the Source Control group for advice. Details: [link]
* Hack IDE can't start because .hhconfig disagrees with your mergebase. Are you in the middle of a rebase? Do you need to arc pull? Once done, restart Hack IDE by clicking in it in the status bar. Details: [link]

The link will be an everpaste that contains enough information for a subject-matter-expert to dig deeper. Often this will be someone from (say) the watchman team.
* A watchman fault stops Hack IDE from working. This is not a bug in Hack or VSCode. Consider looking in the Source Control group for advice.
  * The hg bookmark 'remote/master' is stale.
  *  `hg log -r 'ancestor(.,remote/master')` = COMMIT
  * `hg log -r 'last(public && ::.)' = COMMIT
  * They should be the same."
* Hack IDE can't start because .hhconfig disagrees with your mergebase. Are you in the middle of a rebase? Do you need to arc pull? Once done, restart Hack IDE by clicking in it in the status bar.
  * Mergebase `hg log -r ...` = COMMIT, TIMESTAMP
  * Hash of .hhconfig is '%s'
  * XDB has no record of that hash at that timestamp
  * but it does see the hash at other timestamps: ..."

I've renamed `Saved_state_loader.log_string_of_error` to `.debug_details_of_error`. That's to make it clear that the consumer of this API has authority to decide what goes into the log. The consumer can put short message, medium message, long message, or debug details into the log, or any combination. It makes clear that the `.debug_details` API will *not* include `long_message`, for instance. (that's done in a future diff)

The change will be a multi-diff stack. This diff's job is to expose the API from saved_state_loader. A future diff will change it so short+medium+long are correctly exposed over LSP.

Reviewed By: arxanas

Differential Revision: D20480785

fbshipit-source-id: b0432accc308da31a475f2f2563ca6d140f789b1